pyproject.toml
[project]
name = "heidi"
version = "0.1.0"
description = "Clinical Decision Support System with multi-server MCP architecture"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43.1",
    "beautifulsoup4>=4.14.2",
    "deep-translator>=1.11.0",
    "elevenlabs>=1.0.0",
    "fastapi>=0.121.2",
    "fastmcp>=2.13.0.2",
    "firebase-admin>=7.1.0",
    "gtts>=2.5.4",
    "httpx>=0.27.0",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "mcp>=1.0.0",
    "numpy>=1.26.0,<2.0",
    "pillow-heif>=1.1.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.0",
    "reportlab>=4.4.7",
    "requests>=2.32.5",
    "resend>=2.0.0",
    "supabase>=2.0.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "vcrpy>=6.0.0",
    "pytest-recording>=0.13.0",
    "pytest-xdist>=3.5.0",
    "respx>=0.21.0",
    "pytest-timeout>=2.3.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "vcr: mark test to use VCR cassette recording",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
    "llm: marks tests that require LLM calls",
    "external_service: marks tests hitting external services",
]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "-ra",
]

[tool.coverage.run]
source = ["api", "servers"]
branch = true
omit = [
    "tests/*",
    "*/__pycache__/*",
    ".venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
